Consolidate login helpers into the login_manager
login_manager is a subpackage (not a module) containing
- the 'manager' module
- tokenstore
- auth flow helpers + local server
- an '_old_config' module (just for removal of old config)
