+++ src/globus_cli/helpers/__init__.py
@@ -1,19 +1,5 @@
-from globus_cli.helpers.auth_flows import do_link_auth_flow, do_local_server_auth_flow
-from globus_cli.helpers.delegate_proxy import (
-    fill_delegate_proxy_activation_requirements,
-)
-from globus_cli.helpers.local_server import (
-    LocalServerError,
-    is_remote_session,
-    start_local_server,
-)
+from .delegate_proxy import fill_delegate_proxy_activation_requirements
 
 __all__ = [
-    "print_version",
-    "start_local_server",
-    "is_remote_session",
-    "LocalServerError",
     "fill_delegate_proxy_activation_requirements",
-    "do_link_auth_flow",
-    "do_local_server_auth_flow",
 ]
